Revert "BitcoindGetBlockResponse: Change nonce type to long"
This reverts commit c70ccb1685c108290f8ee3ce9c71d77c9e4c9f00.

+++ bitcoind/src/main/java/bisq/wallets/bitcoind/rpc/responses/BitcoindGetBlockResponse.java
@@ -23,7 +23,7 @@ public class BitcoindGetBlockResponse<T> extends JsonRpcResponse<BitcoindGetBloc
         private long time;
         @Json(name = "mediantime")
         private long medianTime;
-        private long nonce;
+        private int nonce;
         private String bits;
 
         private String target;
